new design

+++ Web/WebEminari.Web/Controllers/WebEminarsController.cs
@@ -5,6 +5,7 @@
     using System.IO;
     using System.Linq;
     using System.Security.Claims;
+    using System.Text;
     using System.Threading.Tasks;
 
     using Microsoft.AspNetCore.Authorization;
@@ -16,12 +17,13 @@
     using WebEminari.Data;
     using WebEminari.Data.Models;
     using WebEminari.Services.Data;
+    using WebEminari.Services.Messaging;
     using WebEminari.Web.ViewModels.WebEminars;
 
-    public class WebEminarsController : Controller
+    public class WebEminarsController : BaseController
     {
-        private const int ItemsPerPage = 6;
-
+        private const int ItemsPerPage = 10;
+        private readonly IEmailGridSender emailSender;
         private readonly ICategoriesService categoriesService;
 
         public IWebHostEnvironment WebHostEnviroment { get; }
@@ -31,21 +33,27 @@
         private readonly IWebEminarsService webEminarService;
 
         private readonly ApplicationDbContext context;
+         private readonly IInformationService informationService;
 
-        public WebEminarsController(ICategoriesService categoriesService, IWebEminarsService webEminarService, ApplicationDbContext context, UserManager<ApplicationUser> userManager, IWebHostEnvironment webHostEnviroment)
+        public WebEminarsController(IInformationService informationService, IEmailGridSender emailSender, ICategoriesService categoriesService, IWebEminarsService webEminarService, ApplicationDbContext context, UserManager<ApplicationUser> userManager, IWebHostEnvironment webHostEnviroment)
         {
             this.userManager = userManager;
             this.context = context;
             this.WebHostEnviroment = webHostEnviroment;
             this.webEminarService = webEminarService;
+            this.emailSender = emailSender;
             this.categoriesService = categoriesService;
+            this.informationService = informationService;
         }
 
 
         [Authorize]
-        public IActionResult Index()
+        public async Task<IActionResult> Index()
         {
-            return this.View("Index");
+
+            var model = await this.informationService.GenerateAdministrationIndexViewModel();
+
+            return this.View(model);
         }
 
         [Authorize]
@@ -64,6 +72,26 @@
             return this.View(items);
         }
 
+        [HttpGet]
+        [Authorize]
+        public async Task<IActionResult> LikedByMe(string stateFilter, int categoryId, string searchString, int id = 1)
+        {
+            var userId = this.GetUserId();
+
+            var viewModel = new WebEminarsListViewModel()
+            {
+                PageNumber = id,
+                WebEminars = this.webEminarService.GetLikedByUserId<WebEminarsInListViewModel>(userId, stateFilter, id, ItemsPerPage, searchString, categoryId),
+                WebEminarsCount = this.webEminarService.GetCount(),
+                ItemsPerPage = ItemsPerPage,
+                SearchText = searchString,
+                StateFilter = stateFilter,
+                CategoriesItems = this.categoriesService.GetAllAsKeyValuePairs(),
+            };
+
+            return this.View(viewModel);
+        }
+
         public IActionResult All(string stateFilter, int categoryId, string searchString, int id = 1)
         {
             if (id <= 0)
@@ -75,6 +103,7 @@
             {
                 PageNumber = id,
                 WebEminars = this.webEminarService.GetAll<WebEminarsInListViewModel>(stateFilter, id, ItemsPerPage, searchString, categoryId),
+                Vebinari = this.webEminarService.GetAllWebinars<WebEminarViewModel>(),
                 WebEminarsCount = this.webEminarService.GetCount(),
                 ItemsPerPage = ItemsPerPage,
                 SearchText = searchString,
@@ -155,8 +184,8 @@
         public IActionResult BookedPeople(int id)
         {
             var webEminar = this.webEminarService.GetById<WebEminarViewModel>(id);
-
-            return this.View(webEminar);
+            var userBookings = context.UserBookings.Where(x => x.WebEminarId == webEminar.Id);
+            return this.View(userBookings);
         }
 
         [Authorize]
@@ -164,7 +193,7 @@
         {
             var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
             var items = this.context.WebEminars;
-            if(items.Any(webEminar => webEminar.AddedByUserId == userId))
+            if(items.Any(webEminar => webEminar.AddedByUserId == userId && webEminar.Id == id))
             {
                 var inputModel = this.webEminarService.GetById<WebEminarViewModel>(id);
 
@@ -204,8 +233,9 @@
         public IActionResult EditWithVideo(int id)
         {
             var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
-            var items = this.context.WebEminars;
-            if (items.Any(webEminar => webEminar.AddedByUserId == userId))
+            var items = this.context.WebEminars.Where(webEminar => webEminar.AddedByUserId == userId);
+
+            if (items.Any(webEminar => webEminar.AddedByUserId == userId && webEminar.Id == id))
             {
                 var inputModel = this.webEminarService.GetById<WebEminarWithVideoViewModel>(id);
 
@@ -235,11 +265,54 @@
             return this.RedirectToAction(nameof(this.Details), new { id });
         }
 
+        [Authorize]
+        public IActionResult IzminalToKachen(int id)
+        {
+            var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var items = this.context.WebEminars;
+            if (items.Any(webEminar => webEminar.AddedByUserId == userId && webEminar.Id == id))
+            {
+                var inputModel = this.webEminarService.GetById<WebEminarWithVideoViewModel>(id);
+
+                inputModel.CategoriesItems = this.categoriesService.GetAllAsKeyValuePairs();
+
+                return this.View(inputModel);
+            }
+            else
+            {
+                throw new ApplicationException();
+            }
+        }
+
+        [Authorize]
+        [HttpPost]
+        public async Task<IActionResult> IzminalToKachen(int id, WebEminarWithVideoViewModel input)
+        {
+            if (!this.ModelState.IsValid)
+            {
+                input.CategoriesItems = this.categoriesService.GetAllAsKeyValuePairs();
+                return this.View(input);
+            }
+
+
+            await this.webEminarService.IzminalToKachen(id, input);
+
+            return this.RedirectToAction(nameof(this.Details), new { id });
+        }
+
         [HttpPost]
         public async Task<IActionResult> Delete(int id)
         {
-            await this.webEminarService.DeleteAsync(id);
-
+            var userId = this.User.FindFirst(ClaimTypes.NameIdentifier).Value;
+            var items = this.context.WebEminars;
+            if (items.Any(webEminar => webEminar.AddedByUserId == userId && webEminar.Id == id))
+            {
+                await this.webEminarService.DeleteAsync(id);
+            }
+            else
+            {
+                throw new ApplicationException();
+            }
             return this.RedirectToAction(nameof(this.All));
         }
 
@@ -266,6 +339,22 @@
         {
             await this.webEminarService.BookEvent(eventId, await this.userManager.GetUserAsync(this.User));
             return this.RedirectToAction("Details", new { id = eventId });
+        }
+
+        [HttpPost]
+        [Authorize]
+        public async Task<IActionResult> SendToEmail(int id, string content)
+        {
+            var user = await this.userManager.GetUserAsync(this.User);
+            var webEminar = this.webEminarService.GetById<WebEminarWithVideoViewModel>(id);
+            var html = new StringBuilder();
+            html.AppendLine($"<h1>{webEminar.Title}</h1>");
+            html.AppendLine($"<h1>От {user.Email}</h1>");
+            html.AppendLine($"<h1>{webEminar.CategoryName}</h1>");
+            html.AppendLine($"<h1>{content}</h1>");
+            html.AppendLine($"<img src=\"{webEminar.ThumbnailUrl}\" />");
+            await this.emailSender.SendEmailAsync("hristomir.hari@gmail.com", "WebEminars", webEminar.AddedByUserEmail, webEminar.Title, html.ToString());
+            return this.RedirectToAction(nameof(this.Details), new { id });
         }
     }
 }
